+++ client/my-sites/email/gsuite/gsuite-purchase-cta/test/index.js
